+++ jdbc/src/main/java/org/springframework/transaction/TransactionExecutor.java
@@ -1,0 +1,54 @@
+package org.springframework.transaction;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+import java.util.function.Supplier;
+import javax.sql.DataSource;
+import org.springframework.dao.DataAccessException;
+import org.springframework.jdbc.datasource.DataSourceUtils;
+
+public class TransactionExecutor {
+
+    private final DataSource dataSource;
+
+    public TransactionExecutor(final DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
+
+    public <T> T execute(final Supplier<T> supplier) {
+        final Connection connection = DataSourceUtils.getConnection(dataSource);
+        try {
+            connection.setAutoCommit(false);
+            final T t = supplier.get();
+            connection.commit();
+            return t;
+        } catch (SQLException | DataAccessException e) {
+            rollback(connection);
+            throw new DataAccessException(e);
+        } finally {
+            DataSourceUtils.releaseConnection(connection, dataSource);
+        }
+    }
+
+    public void execute(final Runnable runnable) {
+        final Connection connection = DataSourceUtils.getConnection(dataSource);
+        try {
+            connection.setAutoCommit(false);
+            runnable.run();
+            connection.commit();
+        } catch (SQLException | DataAccessException e) {
+            rollback(connection);
+            throw new DataAccessException(e);
+        } finally {
+            DataSourceUtils.releaseConnection(connection, dataSource);
+        }
+    }
+
+    private void rollback(final Connection connection) {
+        try {
+            connection.rollback();
+        } catch (SQLException e) {
+            throw new DataAccessException();
+        }
+    }
+}
